Adding Django-tagging-ng library.
It's based on the django-tagging library, with
additional support for multi-lingual and
synonyms, hierarchical, etc...

Also note there's a SQL file in the project root that will
add the column to your breakout_venue table. For ease of use
with the admin interface, you can setup a field on any model
you want to use tags with the TagField() model field and
django-tagging will "do the right thing" behind the scenes.