Importing JQuery 1.3.2 Importing JQuery lightbox and UI.
Added template tags to test for whether or not we have a
venues image, and if so display it. We should look at using
a proper thumbnailing package like thumbs.py or photologue.

Right now I'm just hardcoding the width/height attribs of
the img tag.

Also added lightbox effect on small venue images on the
events listing page.